fix: 리뷰 코멘트 수정
